Clean code, make it more object
