+++ src/mpkiservice/main.py
@@ -5,16 +5,13 @@ from Crypto.Cipher import AES
 from fastapi import Depends, FastAPI, HTTPException, status
 from fastapi.security import HTTPBasic, HTTPBasicCredentials
 from passlib.apache import HtpasswdFile
-from pydantic import SecretStr
 
 from .config import settings
 from .mpki import (
+    Authority,
     Certificate,
     Localisation,
     Partner,
-    create_certificate,
-    find_certificate_index,
-    revoke_certificate,
 )
 
 _logger = logging.getLogger(__name__)
@@ -58,13 +55,17 @@ def get_current_passphrase(
     return passphrase.decode("utf8")
 
 
-def get_current_org(credentials: HTTPBasicCredentials = Depends(security)):
-    return get_current_credentials(credentials).username
+def get_current_authority(credentials: HTTPBasicCredentials = Depends(security)):
+    org = get_current_credentials(credentials).username
+    return Authority(settings, org)
 
 
 @app.get("/certs/{serial}")
-async def get_cert(serial: str, org: str = Depends(get_current_org)):
-    certificate = find_certificate_index(org, serial)
+async def get_cert(
+    serial: str,
+    authority: Authority = Depends(get_current_authority),
+):
+    certificate = authority.find_certificate_index(serial)
     if not certificate:
         raise HTTPException(status_code=404, detail="Certificate not found")
     return certificate
@@ -75,17 +76,16 @@ async def create_cert(
     certificate: Certificate,
     partner: Partner,
     location: Localisation,
-    org: str = Depends(get_current_org),
-    passphrase: SecretStr = Depends(get_current_passphrase),
+    authority: Authority = Depends(get_current_authority),
+    passphrase: str = Depends(get_current_passphrase),
 ):
-    return create_certificate(org, certificate, partner, location, passphrase)
+    return authority.create_certificate(certificate, partner, location, passphrase)
 
 
 @app.delete("/certs/{serial}")
 async def revoke_cert(
     serial: str,
+    authority: Authority = Depends(get_current_authority),
     passphrase: str = Depends(get_current_passphrase),
-    org: str = Depends(get_current_org),
 ):
-    certificate = find_certificate_index(org, serial)
-    return revoke_certificate(org, certificate, passphrase)
+    return authority.revoke_certificate(serial, passphrase)
